refactor: Change MainActivity to FragmentActivity
