+++ app/src/main/java/com/samrraa/qurioapp/MainActivity.kt
@@ -1,19 +1,13 @@
 package com.samrraa.qurioapp
 
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.samrraa.qurioapp.databinding.ActivityMainBinding
+import androidx.fragment.app.FragmentActivity
 
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
+class MainActivity : FragmentActivity() {
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        installSplashScreen()
-        enableEdgeToEdge()
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
+        setContentView(R.layout.activity_main)
 
         (application as QurioApp).appComponent.inject(this)
     }
